Move harp qc tests for to its own module
