+++ src/contraqctor/qc/harp/__init__.py
@@ -1,0 +1,9 @@
+from .harp_device import HarpDeviceTestSuite, HarpDeviceTypeTestSuite, HarpHubTestSuite
+from .sniff_detector import HarpSniffDetectorTestSuite
+
+__all__ = [
+    "HarpDeviceTestSuite",
+    "HarpDeviceTypeTestSuite",
+    "HarpHubTestSuite",
+    "HarpSniffDetectorTestSuite",
+]
